добавлены проверки на права пользователя, методы в CommentService и CommentController, изменены мапперы, исправлены ошибки, возникающие при авторизации

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -1,26 +1,26 @@
 package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.AdsDto;
 import ru.skypro.homework.dto.CreateAds;
 import ru.skypro.homework.dto.ResponseWrapperAds;
+import ru.skypro.homework.dto.Role;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.User;
 import ru.skypro.homework.exception.ImageNotFoundException;
 import ru.skypro.homework.exception.AdsNotFoundException;
 import ru.skypro.homework.exception.UserNotFoundException;
-import ru.skypro.homework.mapper.AdsMapper;
-import ru.skypro.homework.mapper.CommentMapper;
+import ru.skypro.homework.mapper.AdsMapperInterface;
 import ru.skypro.homework.repository.AdsRepository;
-import ru.skypro.homework.repository.CommentRepository;
+import ru.skypro.homework.repository.ImageRepository;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.ImageService;
-import ru.skypro.homework.service.UserService;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -33,24 +33,27 @@ import java.util.stream.Collectors;
 public class AdsServiceImpl implements AdsService {
     private final AdsRepository adsRepository;
     private final ImageService imageService;
-    private final AdsMapper adsMapper;
     private final UserRepository userRepository;
+    private final ImageRepository imageRepository;
+
+
 
     @Override
     public Ads addAd(CreateAds createAds, MultipartFile image, Authentication authentication) throws IOException {
-
-        User user = userRepository.findByEmailIgnoreCase(SecurityContextHolder.getContext()
-                .getAuthentication().getName()).orElseThrow();
-        Ads ads = adsMapper.mapToAds(createAds, image);
+        User user = userRepository.findByEmailIgnoreCase(authentication.getName()).orElseThrow(() ->
+                new UserNotFoundException("Пользователь с e-mail" + authentication.getName() + "не найден"));
+        Ads ads = AdsMapperInterface.INSTANCE.toEntity(createAds);
+        checkPermissionsToWorkWithAds (ads, authentication);
         ads.setAuthor(user);
         ads.setImage(imageService.uploadImage(image));
         return adsRepository.save(ads);
-
     }
 
     @Override
-    public Ads removeAd(Integer id, Authentication authentication ) {
+    public Ads removeAd(Integer id, Authentication authentication) {
         Ads ads = findAdsById(id);
+        checkPermissionsToWorkWithAds (ads, authentication);
+        imageRepository.deleteImageByAdsId(id);
         adsRepository.delete(ads);
         return ads;
     }
@@ -63,15 +66,17 @@ public class AdsServiceImpl implements AdsService {
 
 
     @Override
-    public Ads getAds(Integer id) {
+    public Ads getAds(Integer id, Authentication authentication) {
         Ads ads = findAdsById(id);
-        return  ads;
+        checkPermissionsToWorkWithAds (ads, authentication);
+        return ads;
     }
 
 
     @Override
     public Ads updateAds(Integer id, CreateAds createAds, Authentication authentication) {
         Ads ads = findAdsById(id);
+        checkPermissionsToWorkWithAds (ads, authentication);
         ads.setDescription(createAds.getDescription());
         ads.setPrice(createAds.getPrice());
         ads.setTitle(createAds.getTitle());
@@ -79,18 +84,18 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public ResponseWrapperAds getAdsMe(String username) {
+    public ResponseWrapperAds getAdsMe(Authentication authentication) {
+        String username = authentication.getName();
         User user = userRepository.findByEmailIgnoreCase(username).orElseThrow(() ->
                 new UserNotFoundException("Пользователь с e-mail" + username + "не найден"));
 
-        List <AdsDto> adsDtoList  = adsRepository.findAllByAuthorId(user.getId()).stream().map (x -> adsMapper.mapToAdsDto(x)).
+        List<AdsDto> adsDtoList = adsRepository.findAllByAuthorId(user.getId()).stream().map(x -> AdsMapperInterface.INSTANCE.toDto(x)).
                 collect(Collectors.toList());
-            ResponseWrapperAds responseWrapperAdsDto = new ResponseWrapperAds();
-            responseWrapperAdsDto.setCount(adsDtoList.size());
-            responseWrapperAdsDto.setResults(adsDtoList);
-            return responseWrapperAdsDto;
-        }
-
+        ResponseWrapperAds responseWrapperAdsDto = new ResponseWrapperAds();
+        responseWrapperAdsDto.setCount(adsDtoList.size());
+        responseWrapperAdsDto.setResults(adsDtoList);
+        return responseWrapperAdsDto;
+    }
 
 
     @Override
@@ -99,16 +104,38 @@ public class AdsServiceImpl implements AdsService {
             throw new ImageNotFoundException("Изображение не загружено");
         }
         Ads ads = findAdsById(id);
+        checkPermissionsToWorkWithAds (ads, authentication);
         imageService.removeImage(ads.getImage());
         ads.setImage(imageService.uploadImage(image));
         adsRepository.save(ads);
     }
 
+    @Override
+    public ResponseWrapperAds getAdsByTitleLike(String title) {
+        List<AdsDto> ads = adsRepository.findByTitleContainingIgnoreCase(title).stream()
+                .map(x -> AdsMapperInterface.INSTANCE.toDto(x)).collect(Collectors.toList());
+        ResponseWrapperAds responseWrapperAdsDto = new ResponseWrapperAds();
+        responseWrapperAdsDto.setCount(ads.size());
+        responseWrapperAdsDto.setResults(ads);
+        return responseWrapperAdsDto;
+    }
+
 
     public Ads findAdsById(Integer id) {
+
         return adsRepository.findById(id).orElseThrow(
                 () -> new AdsNotFoundException("Объявление с id " + id + " не найдено"));
     }
+
+    public boolean checkPermissionsToWorkWithAds(Ads ads, Authentication authentication) {
+        String username = authentication.getName();
+        User user = userRepository.findByEmailIgnoreCase(username).orElseThrow(() ->
+                new UserNotFoundException("Пользователь с e-mail" + username + "не найден"));
+        if (!authentication.getAuthorities().contains(Role.ADMIN) && user.getId() != ads.getAuthor().getId()) {
+            throw new AccessDeniedException("Редактировать/удалять объявления может только автор объявления или Админ!");
+        }
+        return true;
+    };
